dev-collab-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==0.28.0
python-dotenv==1.0.0
PyJWT==2.8.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-session==0.4.0

